+++ src/components/layouts/Navbar.tsx
@@ -1,26 +1,99 @@
-import { Home, LineChart } from "lucide-react";
-import Link from "next/link";
 import React from "react";
+import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { useTheme } from "next-themes";
+import { Button } from "@/components/ui/button";
+import {
+  CircleUser,
+  Home,
+  LineChart,
+  Menu,
+  Package2,
+  Search,
+} from "lucide-react";
+import Link from "next/link";
+import { Input } from "../ui/input";
+import { ModeToggle } from "../ModeToggle";
 
-const Navbar = () => {
+const Sidebar = () => {
+  const { resolvedTheme } = useTheme();
   return (
-    <nav className="grid items-start px-2 text-sm font-medium lg:px-4">
-      <Link
-        href="#"
-        className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-      >
-        <Home className="h-4 w-4" />
-        Dashboard
-      </Link>
-      <Link
-        href="#"
-        className="flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary"
-      >
-        <LineChart className="h-4 w-4" />
-        Analytics
-      </Link>
-    </nav>
+    <header className="flex h-14 items-center gap-4 border-b  dark:bg-gray-900 px-4 lg:h-[60px] lg:px-6">
+      <Sheet>
+        <SheetTrigger asChild>
+          <Button variant="outline" size="icon" className="shrink-0 md:hidden">
+            <Menu
+              className={`h-5 w-5 ${
+                resolvedTheme === "dark" ? "text-white" : "text-black"
+              }`}
+            />
+            <span className="sr-only">Toggle navigation menu</span>
+          </Button>
+        </SheetTrigger>
+        <SheetContent side="left" className="flex flex-col bg-white">
+          <nav className="grid gap-2 text-lg font-medium">
+            <Link
+              href="#"
+              className="flex items-center gap-2 text-lg font-semibold"
+            >
+              <Package2 className="h-6 w-6" />
+              <span className="text-gray-400">Dashboard</span>
+            </Link>
+            <Link
+              href="#"
+              className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
+            >
+              <Home className="h-5 w-5" />
+              Dashboard
+            </Link>
+            <Link
+              href="#"
+              className="mx-[-0.65rem] flex items-center gap-4 rounded-xl px-3 py-2 text-muted-foreground hover:text-foreground"
+            >
+              <LineChart className="h-5 w-5" />
+              Analytics
+            </Link>
+          </nav>
+        </SheetContent>
+      </Sheet>
+      <div className="w-full flex-1">
+        <form>
+          <div className="relative">
+            <Search className="absolute left-2.5 top-2.5 h-4 w-4 text-muted-foreground" />
+            <Input
+              type="search"
+              placeholder="Search products..."
+              className="w-full appearance-none bg-background pl-8 shadow-none md:w-2/3 lg:w-1/3"
+            />
+          </div>
+        </form>
+      </div>
+
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button variant="secondary" size="icon" className="rounded-full">
+            <CircleUser className="h-5 w-5" />
+            <span className="sr-only">Toggle user menu</span>
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end">
+          <DropdownMenuLabel>Account</DropdownMenuLabel>
+          <DropdownMenuSeparator />
+          <DropdownMenuItem>Settings</DropdownMenuItem>
+          <DropdownMenuSeparator />
+          <DropdownMenuItem>Logout</DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+      <ModeToggle />
+    </header>
   );
 };
 
-export default Navbar;
+export default Sidebar;
